infrastructure/testharness.js/worker-shared.html

<!DOCTYPE HTML>
<html>
<head>
<meta name="variant" content="">
<meta name="variant" content="?remove-promise">
<title>Example with a shared worker</title>
<script src="resources/variant-promise.js"></script>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
</head>
<body>
<h1>Shared Web Worker Tests</h1>
<p>Demonstrates running <tt>testharness</tt> based tests inside a shared worker.
<p>The test harness should time out due to one of the tests in the worker timing out.
<p>This test assumes that the browser supports <a href="http://www.w3.org/TR/workers/#shared-workers-and-the-sharedworker-interface">shared web workers</a>.
<div id="log"></div>

<script>
test(
    function(t) {
        assert_true("SharedWorker" in self,
                    "Browser should support SharedWorkers");
    },
    "Browser supports SharedWorkers");

fetch_tests_from_worker(new SharedWorker("resources/worker.js",
                                         "My shared worker"));
</script>
</body>
